requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==4.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
jsonfield==3.1.0
numpy==1.21.4
pandas==1.3.5
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
xlrd==2.0.1
xlwt==1.3.0
